Support custom headers for SSR pages and API endpoints
There are some minor differences between Netlify and AWS, such as AWS
having support for multi-value headers. next-aws-lambda is made for AWS,
so we need to resolve those differences to make everything work with
Netlify.

We support headers in SSR pages and API endpoints by converting the
first entry in each multi-value header into a plain header.

Netlify also requires all header values to be in string format, so we
take care of that, too.